airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
apache-airflow-providers-snowflake
apache-airflow-providers-airbyte
pandas
minio
pyarrow
